requirements.txt
requests
dnspython
shodan
reportlab
python-dotenv